Corrected opening sentence in run scripts

+++ run-client.sh
@@ -5,12 +5,12 @@ if [[ $1 = "18" ]]; then
   alias java=java18
 fi
 
-[[ -f "jars/ClientApp.jar" ]] &&
-(
+if [[ -f "jars/ClientApp.jar" ]]; then
   echo "Запуск клиента...";
   java -Xmx4g -XX:OnOutOfMemoryError="kill -9 %p; sh run-client.sh" -jar jars/ClientApp.jar config/client-config.json
-) ||
+else
   echo "Сначала соберите проект: выполните build.sh";
+fi
 
 if [[ $1 = "18" ]]; then
   unalias java
